Rename dog files, seperate tracking logic to shared file
